factor leet function up, unify probe naming
